pyproject.toml
[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

# --------------------------------------------------------------------------- #
# Core project metadata (PEP 621)
# --------------------------------------------------------------------------- #
[project]
name            = "flujo"
version         = "0.2.0" # Version bumped for new features
description     = "Production-ready orchestration for AI agents, built with Pydantic."
readme          = "README.md"
requires-python = ">=3.11,<4.0"

authors = [
  { name = "Alvaro", email = "alvaro@example.com" }
]

license = { text = "MIT" }

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
  "Typing :: Typed",
]

# Runtime dependencies
dependencies = [
  "pydantic>=2.7",
  "pydantic-ai>=0.2.16",
  "pydantic-settings",
  "python-dotenv>=1.0",
  "pyyaml>=6.0",
  "tenacity>=8.2",
  "typer[all]>=0.12",
  "rich>=13.7",
  "logfire>=0.3",
  "sqlvalidator>=0.0.8",
  "pydantic-evals>=0.0.47", # Added for Intelligent Evals
]

# Optional dependency groups (install with: `pip install .[dev]`, etc.)
[project.optional-dependencies]
dev = [
  "ruff",
  "mypy",
  "pip-audit",
  "pytest",
  "pytest-cov",
  "pytest-asyncio[mode-auto]",
  "hypothesis",
  "vcrpy",
  "rich",
  "pytest-benchmark",
  "build",     # Added for package building
  "twine",     # Added for package uploading
]
docs = [
    "mkdocs",
    "mkdocs-material", # Added for a professional theme
    "mkdocstrings[python]",
]
opentelemetry = ["opentelemetry-sdk>=1.26"]
bench = ["numpy"]

# CLI entry point (install adds `flujo` command to PATH)
[project.scripts]
flujo = "flujo.cli.main:app"

# Project links (PEP 621)
[project.urls]
Homepage = "https://github.com/aandresalvarez/rloop"
Repository = "https://github.com/aandresalvarez/rloop"
Documentation = "https://aandresalvarez.github.io/rloop/"
Bug_Tracker = "https://github.com/aandresalvarez/rloop/issues"
Changelog = "https://github.com/aandresalvarez/rloop/blob/main/CHANGELOG.md"

# --------------------------------------------------------------------------- #
# Tool-specific configuration
# --------------------------------------------------------------------------- #
[tool.mypy]
python_version = "3.11"
strict         = true

# Focus type checking on the core library first. Examples are excluded as they are
# meant to be educational and may not follow strict typing.
exclude = [
    "examples/.*",
]

# Allow third-party libraries that do not ship type information.
[[tool.mypy.overrides]]
module = ["yaml", "sqlvalidator", "vcr"]
ignore_missing_imports = true

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
  "e2e: marks tests as end-to-end (requires network access or VCR cassettes)",
  "benchmark: marks tests as benchmarks (requires pytest-benchmark)",
]

# --------------------------------------------------------------------------- #
# Build system configuration
# --------------------------------------------------------------------------- #
[tool.hatch.build.targets.wheel]
packages = ["flujo"]

[tool.hatch.build.targets.sdist]
include = [
    "flujo",
    "tests",
    "docs",
    "examples",
    "README.md",
    "LICENSE",
    "CHANGELOG.md",
    "pyproject.toml",
]

[tool.hatch.build.targets.wheel.shared-data]
"README.md" = "share/doc/flujo/README.md"
"LICENSE" = "share/doc/flujo/LICENSE"
"CHANGELOG.md" = "share/doc/flujo/CHANGELOG.md"
